refactor: move plot function out of segal
